+++ shared/shared/src/main/scala/sc4pac/shared.scala
@@ -12,6 +12,7 @@ final case class BareModule(group: Organization, name: ModuleName) extends BareD
 }
 object BareModule {
   val pkgMarkdownRegex = """`pkg=([^`:\s]+):([^`:\s]+)`""".r
+  given lexOrdering: Ordering[BareModule] = Ordering.by(module => (module.group.value, module.name.value))
 }
 final case class BareAsset(assetId: ModuleName) extends BareDep {
   def orgName = s"${JsonRepoUtil.sc4pacAssetOrg.value}:${assetId.value}"
@@ -25,6 +26,13 @@ object JsonRepoUtil {
       case a: BareAsset => (sc4pacAssetOrg.value, a.assetId.value)
     }
     s"metadata/${group}/${name}/${version}/pkg.json"
+  }
+
+  def extPackageSubPath(dep: BareDep): String = {
+    dep match {
+      case m: BareModule => s"metadata/_extpkg/${m.group.value}/${m.name.value}/latest/pkg.json"
+      case a: BareAsset => s"metadata/_extasset/${a.assetId.value}/latest/pkg.json"
+    }
   }
 
   val sc4pacAssetOrg = Organization("sc4pacAsset")
@@ -47,6 +55,9 @@ abstract class SharedData {
   protected def emptyChecksum: Checksum
 
   implicit val bareModuleRw: ReadWriter[BareModule]
+
+  type Uri
+  implicit val uriRw: ReadWriter[Uri]
 
   case class Dependency(group: String, name: String, version: String) derives ReadWriter
 
@@ -136,7 +147,9 @@ abstract class SharedData {
     info: Info = Info.empty,
     variants: Seq[VariantData],  // should be non-empty, but can consist of a single empty variant
     variantDescriptions: Map[String, Map[String, String]] = Map.empty,  // variantKey -> variantValue -> description
-    metadataSource: Option[SubPath] = None  // path to yaml file
+    metadataSource: Option[SubPath] = None,  // path to yaml file
+    metadataSourceUrl: Option[Uri] = None,  // full URL to yaml file
+    channelLabel: Option[String] = None,
   ) extends PackageAsset {
 
     def toBareDep: BareModule = BareModule(Organization(group), ModuleName(name))
@@ -165,6 +178,23 @@ abstract class SharedData {
     val empty = Info()
   }
 
+  case class ExternalPackage(
+    group: String,
+    name: String,
+    // channel: Option[String],
+    requiredBy: Seq[BareModule],
+  ) derives ReadWriter {
+    def toBareDep: BareModule = BareModule(Organization(group), ModuleName(name))
+  }
+
+  case class ExternalAsset(
+    assetId: String,
+    // channel: Option[String],
+    requiredBy: Seq[BareModule],
+  ) derives ReadWriter {
+    def toBareDep: BareAsset = BareAsset(ModuleName(assetId))
+  }
+
   case class ChannelItem(
     group: String,
     name: String,
@@ -180,8 +210,11 @@ abstract class SharedData {
 
   case class Channel(
     scheme: Int,
+    info: Channel.Info = Channel.Info.empty,
     stats: Channel.Stats = null,  // added between scheme 4 and 5 (backward compatible: recomputed for smaller schemes, so never actually null)
     contents: Seq[ChannelItem],
+    externalPackages: Seq[Channel.ExtPkg] = Seq.empty,  // default for backward compatibility
+    externalAssets: Seq[Channel.ExtAsset] = Seq.empty,  // default for backward compatibility
   ) {
     lazy val versions: Map[BareDep, Seq[(String, Checksum)]] =
       contents.iterator.map(item => item.toBareDep -> item.versions.map(v => v -> item.checksums.getOrElse(v, emptyChecksum))).toMap
@@ -190,19 +223,31 @@ abstract class SharedData {
 
     private val channelRwDefault: ReadWriter[Channel] = macroRW
     implicit val channelRw: ReadWriter[Channel] =
-      channelRwDefault.bimap[Channel](identity, c => if (c.stats != null) c else createAddStats(c.scheme, c.contents))
+      channelRwDefault.bimap[Channel](identity, c => if (c.stats != null) c else createAddStats(c.scheme, c.info, c.contents, c.externalPackages, c.externalAssets))
 
     /* recomputes the channel stats */
-    def createAddStats(scheme: Int, contents: Seq[ChannelItem]): Channel = {
+    def createAddStats(
+      scheme: Int,
+      info: Channel.Info,
+      contents: Seq[ChannelItem],
+      externalPackages: Seq[Channel.ExtPkg],
+      externalAssets: Seq[Channel.ExtAsset],
+    ): Channel = {
       val m = collection.mutable.Map.empty[String, Int]
       for (item <- contents; cat <- item.category) {
         m(cat) = m.getOrElse(cat, 0) + 1
       }
-      Channel(scheme, Stats.fromMap(m), contents = contents)
+      Channel(scheme, info, Stats.fromMap(m), contents = contents, externalPackages, externalAssets)
     }
 
-    def create(scheme: Int, channelData: Iterable[(BareDep, Iterable[(String, PackageAsset, Checksum)])]): Channel = {  // name -> (version, json, sha)
-      createAddStats(scheme, contents = channelData.iterator.collect {
+    def create(
+      scheme: Int,
+      info: Channel.Info,
+      channelData: Iterable[(BareDep, Iterable[(String, PackageAsset, Checksum)])],  // name -> (version, json, sha)
+      externalPackages: Seq[Channel.ExtPkg],
+      externalAssets: Seq[Channel.ExtAsset],
+    ): Channel = {
+      createAddStats(scheme, info, contents = channelData.iterator.collect {
         case (dep, versions) if versions.nonEmpty =>
           val (g, n) = dep match {
             case m: BareModule => (m.group.value, m.name.value)
@@ -218,7 +263,7 @@ abstract class SharedData {
             summary = summaryOpt.getOrElse(""),
             category = catOpt,
           )
-      }.toSeq)
+      }.toSeq, externalPackages, externalAssets)
     }
 
     case class CategoryItem(category: String, count: Int) derives ReadWriter
@@ -241,6 +286,21 @@ abstract class SharedData {
         Stats.fromMap(m)
       }
     }
+
+    case class Info(
+      channelLabel: Option[String],
+      metadataSourceUrl: Option[Uri],
+    ) derives ReadWriter
+    object Info {
+      val empty = Info(None, None)
+    }
+
+    case class ExtPkg(group: String, name: String, checksum: Checksum) derives ReadWriter {
+      def toBareDep: BareModule = BareModule(Organization(group), ModuleName(name))
+    }
+
+    case class ExtAsset(assetId: String, checksum: Checksum) derives ReadWriter
+
   }
 
 }