+++ src/main/scala/sc4pac/Data.scala
@@ -18,6 +18,7 @@ object JsonData extends SharedData {
 
   override type Instant = java.time.Instant
   override type SubPath = os.SubPath
+  override type Uri = java.net.URI
 
   // We use OffsetDateTime.parse instead of Instant.parse for compatibility with Java 8 to 11
   implicit val instantRw: ReadWriter[java.time.Instant] =
@@ -30,7 +31,7 @@ object JsonData extends SharedData {
   implicit val uriRw: ReadWriter[java.net.URI] = readwriter[String].bimap[java.net.URI](_.toString(),
     MetadataRepository.parseChannelUrl(_).left.map(new IllegalArgumentException(_)).toTry.get)
 
-  private def bareModuleRead(s: String) =
+  private[sc4pac] def bareModuleRead(s: String) =
     Sc4pac.parseModule(s) match {
       case Right(mod) => mod
       case Left(err) => throw new IllegalArgumentException(err)