+++ src/main/scala/sc4pac/MetadataRepository.scala
@@ -4,7 +4,7 @@ package sc4pac
 import coursier.core.{Module, Versions, Version}
 import coursier.cache as CC
 import upickle.default.Reader
-import zio.{ZIO, IO}
+import zio.{ZIO, IO, UIO}
 
 import sc4pac.error.*
 import sc4pac.JsonData as JD
@@ -46,6 +46,8 @@ sealed abstract class MetadataRepository(val baseUri: java.net.URI) {
     */
   def fetchModuleJson[A <: JD.PackageAsset : Reader](module: Module, version: String, fetch: MetadataRepository.Fetch): zio.Task[A]
 
+  def fetchExternalPackage(module: BareModule, fetch: MetadataRepository.Fetch): zio.Task[Option[JD.ExternalPackage]]
+
   def iterateChannelContents: Iterator[JD.ChannelItem]
 }
 
@@ -74,13 +76,12 @@ object MetadataRepository {
 
   def create(channelContentsFile: os.Path, baseUri: java.net.URI): IO[ErrStr, MetadataRepository] = {
     if (baseUri.getPath.endsWith(".yaml")) {  // yaml repository
-      for {
-        contents <- ChannelUtil.readAndParsePkgData(channelContentsFile, root = None)
-      } yield {
-        val channelData: Map[BareDep, Map[String, JD.PackageAsset]] =
-          contents.groupMap(_.toBareDep)(data => data.version -> data).view.mapValues(_.toMap).toMap
-        new YamlRepository(baseUri, channelData)
-      }
+      (for {
+        packages <- ChannelUtil.readAndParsePkgData(channelContentsFile, root = None)
+        // Internally, this picks the latest scheme version as we do not have any other input to work with.
+        // If the scheme has been updated, then the yaml files have probably already failed to parse.
+        yamlRepo <- YamlChannelBuilder().resultToRepository(baseUri, packages)
+      } yield yamlRepo).provideSomeLayer(zio.ZLayer.succeed(JD.Channel.Info.empty))
     } else {  // json repository
       val contentsUrl = channelContentsUrl(baseUri).toString
       for {
@@ -109,6 +110,10 @@ object MetadataRepository {
   def latestSubPath(group: String, name: String): os.SubPath = {
     os.SubPath(s"metadata/$group/$name/latest")
   }
+
+  def extPkgJsonSubPath(dep: BareDep): os.SubPath = {
+    os.SubPath(JsonRepoUtil.extPackageSubPath(dep))
+  }
 }
 
 /** This repository operates on a hierarchy of JSON files. The JSON files are loaded on demand.
@@ -117,6 +122,10 @@ private class JsonRepository(
   baseUri: java.net.URI,
   val channel: JD.Channel,
 ) extends MetadataRepository(baseUri) {
+
+  /* the Map of references to external packages stored in the channel contents file */
+  private lazy val externalPackages: Map[BareModule, JD.Channel.ExtPkg] =
+    channel.externalPackages.view.map(item => item.toBareDep -> item).toMap
 
   /** Obtain the raw version strings of `dep` contained in this repository. */
   def getRawVersions(dep: BareDep): Seq[String] = channel.versions.getOrElse(dep, Seq.empty).map(_._1)
@@ -150,6 +159,18 @@ private class JsonRepository(
           .flatMap((jsonStr: String) => JsonIo.read[A](jsonStr, errMsg = remoteUrl))
   }
 
+  def fetchExternalPackage(module: BareModule, fetch: MetadataRepository.Fetch): zio.Task[Option[JD.ExternalPackage]] = {
+    externalPackages.get(module) match {
+      case None => ZIO.succeed(None)
+      case Some(extPkg) =>
+        val remoteUrl = baseUri.resolve(MetadataRepository.extPkgJsonSubPath(module).segments0.mkString("/")).toString
+        val jsonArtifact = Artifact(remoteUrl, changing = false, checksum = extPkg.checksum)
+        fetch(jsonArtifact)
+          .flatMap((jsonStr: String) => JsonIo.read[JD.ExternalPackage](jsonStr, errMsg = remoteUrl))
+          .map(Some(_))
+    }
+  }
+
   def iterateChannelContents: Iterator[JD.ChannelItem] = channel.contents.iterator
 }
 
@@ -157,7 +178,10 @@ private class JsonRepository(
   */
 private class YamlRepository(
   baseUri: java.net.URI,
-  channelData: Map[BareDep, Map[String, JD.PackageAsset]]  // name -> version -> json
+  val channel: JD.Channel,
+  channelData: collection.Map[BareDep, Map[String, JD.PackageAsset]],  // name -> version -> json
+  externalPackages: collection.Map[BareModule, JD.ExternalPackage],
+  externalAssets: collection.Map[BareAsset, JD.ExternalAsset],
 ) extends MetadataRepository(baseUri) {
 
   def getRawVersions(dep: BareDep): Seq[String] = {
@@ -175,13 +199,45 @@ private class YamlRepository(
     }
   }
 
-  // We pick the latest scheme version as we do not have any other input to work with.
-  // If the scheme has been updated, then the yaml files have probably already failed to parse.
-  // TODO Refactor to remove channelData in favor of channel.
-  lazy val channel = JD.Channel.create(
-    scheme = Constants.channelSchemeVersions.max,
-    channelData.view.mapValues(_.view.map { case (version, pkgData) => (version, pkgData, JD.Checksum.empty) })
-  )
+  def fetchExternalPackage(module: BareModule, fetch: MetadataRepository.Fetch): zio.Task[Option[JD.ExternalPackage]] = {
+    ZIO.succeed(externalPackages.get(module))
+  }
 
   def iterateChannelContents: Iterator[JD.ChannelItem] = channel.contents.iterator
+}
+
+private class YamlChannelBuilder extends ChannelBuilder[Nothing] {
+  import scala.jdk.CollectionConverters.*
+  private val contents = (new java.util.concurrent.ConcurrentHashMap[BareDep, Map[String, JD.PackageAsset]]()).asScala  // name -> version -> json
+  private val extPkgs = (new java.util.concurrent.ConcurrentHashMap[BareModule, JD.ExternalPackage]()).asScala
+  private val extAssets = (new java.util.concurrent.ConcurrentHashMap[BareAsset, JD.ExternalAsset]()).asScala
+
+  def storePackage(data: JD.Package): UIO[JD.Checksum] = ZIO.succeed {
+    contents.updateWith(data.toBareDep) { versionsMapOpt => Some {
+      versionsMapOpt.getOrElse(Map.empty) + (data.version -> data)
+    }}
+    JD.Checksum.empty
+  }
+
+  def storeAsset(data: JD.Asset): UIO[JD.Checksum] = ZIO.succeed {
+    contents.updateWith(data.toBareDep) { versionsMapOpt => Some {
+      versionsMapOpt.getOrElse(Map.empty) + (data.version -> data)
+    }}
+    JD.Checksum.empty
+  }
+
+  def storeExtPackage(data: JD.ExternalPackage): UIO[JD.Checksum] = ZIO.succeed {
+    extPkgs += data.toBareDep -> data
+    JD.Checksum.empty
+  }
+
+  def storeExtAsset(data: JD.ExternalAsset): UIO[JD.Checksum] = ZIO.succeed {
+    extAssets += data.toBareDep -> data
+    JD.Checksum.empty
+  }
+
+  def resultToRepository(baseUri: java.net.URI, packages: Seq[JD.PackageAsset]): zio.URIO[JD.Channel.Info, YamlRepository] =
+    result(packages).map { channel =>
+      YamlRepository(baseUri, channel, contents, extPkgs, extAssets)
+    }
 }